+++ HeFeiSaledServices/SaledServices/Store/StoreHouseInnerNGForm.cs
@@ -231,32 +231,36 @@ namespace SaledServices
 
         private void delete_Click(object sender, EventArgs e)
         {
-            try
-            {
-                SqlConnection conn = new SqlConnection(Constlist.ConStr);
-                conn.Open();
+             DialogResult dr = MessageBox.Show("你确定删除吗？", "确认对话框", MessageBoxButtons.OKCancel, MessageBoxIcon.Question);
+             if (dr == DialogResult.OK)
+             {
+                 try
+                 {
+                     SqlConnection conn = new SqlConnection(Constlist.ConStr);
+                     conn.Open();
 
-                if (conn.State == ConnectionState.Open)
-                {
-                    SqlCommand cmd = new SqlCommand();
-                    cmd.Connection = conn;
-                    cmd.CommandText = "Delete from " + tableName + " where id = " + dataGridView1.SelectedCells[0].Value.ToString();
-                    cmd.CommandType = CommandType.Text;
-                    cmd.ExecuteNonQuery();
-                }
-                else
-                {
-                    MessageBox.Show("SaledService is not opened");
-                }
+                     if (conn.State == ConnectionState.Open)
+                     {
+                         SqlCommand cmd = new SqlCommand();
+                         cmd.Connection = conn;
+                         cmd.CommandText = "Delete from " + tableName + " where id = " + dataGridView1.SelectedCells[0].Value.ToString();
+                         cmd.CommandType = CommandType.Text;
+                         cmd.ExecuteNonQuery();
+                     }
+                     else
+                     {
+                         MessageBox.Show("SaledService is not opened");
+                     }
 
-                conn.Close();
-                MessageBox.Show("删除完毕!");
-                query_Click(null, null);
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.ToString());
-            }
+                     conn.Close();
+                     MessageBox.Show("删除完毕!");
+                     query_Click(null, null);
+                 }
+                 catch (Exception ex)
+                 {
+                     MessageBox.Show(ex.ToString());
+                 }
+             }
         }
 
         private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)
